Standardize Steps section wording and style
The Digsite Pendant is already the only sane way of traveling to
Fossil Island, so it doesn't need to be strong.

This change has the added benefit of emphasizing the location over
the item. A pendant charge can be wasted if thinking "Digsite" as
the name of the item vs. "Digsite" the location.